+++ src/main/java/com/hencjo/summer/security/HttpBasicAuthenticator.java
@@ -1,23 +1,21 @@
 package com.hencjo.summer.security;
 
-import java.io.IOException;
-import java.util.Optional;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import com.hencjo.summer.security.api.Authenticator;
-import com.hencjo.summer.security.api.Authenticator2;
+import com.hencjo.summer.security.api.CredentialsAuthenticator;
+import com.hencjo.summer.security.api.CredentialsAuthenticator.Credentials;
 import com.hencjo.summer.security.api.RequestMatcher;
 import com.hencjo.summer.security.api.Responder;
-import com.hencjo.summer.security.utils.Base64;
 import com.hencjo.summer.security.utils.Charsets;
 
-public final class HttpBasicAuthenticator<T> {
-	private final Authenticator2 authenticator;
-	private final SummerAuthenticatedUser summerAuthenticatedUser = new SummerAuthenticatedUser();
-	private final Base64 base64 = new Base64();
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.Optional;
+
+public final class HttpBasicAuthenticator {
+	private final CredentialsAuthenticator authenticator;
 	private final String realm;
 
-	public HttpBasicAuthenticator(Authenticator2 authenticator, String realm) {
+	public HttpBasicAuthenticator(CredentialsAuthenticator authenticator, String realm) {
 		if (realm == null) throw new IllegalArgumentException("Realm may not be null.");
 		if (realm.contains("\"")) throw new IllegalArgumentException("Realm may not contain quotes (\")");
 		this.authenticator = authenticator;
@@ -28,18 +26,7 @@ public final class HttpBasicAuthenticator<T> {
 		return new RequestMatcher() {
 			@Override
 			public boolean matches(HttpServletRequest request) {
-				String authorization = request.getHeader("Authorization");
-				if (authorization == null) return false;
-				if (!authorization.startsWith("Basic")) return false;
-				String usernameColonPassword = new String(base64.decode(authorization.substring(6)), Charsets.UTF8);
-				String[] split = usernameColonPassword.split(":");
-				if (split.length != 2) return false;
-				String username = split[0];
-				String password = split[1];
-				Optional<String> authenticate = authenticator.authenticate(username, password);
-				if (!authenticate.isPresent()) return false;
-				summerAuthenticatedUser.set(request, authenticate.get());
-				return true;
+				return hasAuthentication(request) != null;
 			}
 			
 			@Override
@@ -47,6 +34,26 @@ public final class HttpBasicAuthenticator<T> {
 				return "HttpBasic";
 			}
 		};
+	}
+
+	public String hasAuthentication(HttpServletRequest request) {
+		Credentials credentials = credentials(request).orElse(null);
+		if (credentials == null) return null;
+
+		Optional<String> authenticate = authenticator.authenticate(credentials);
+		return authenticate.orElse(null);
+	}
+
+	public static Optional<Credentials> credentials(HttpServletRequest request) {
+		String authorization = request.getHeader("Authorization");
+		if (authorization == null) return Optional.empty();
+		if (!authorization.startsWith("Basic")) return Optional.empty();
+		String usernameColonPassword = new String(java.util.Base64.getDecoder().decode(authorization.substring(6)), Charsets.UTF8);
+		String[] split = usernameColonPassword.split(":");
+		if (split.length != 2) return Optional.empty();
+		String username = split[0];
+		String password = split[1];
+		return Optional.of(new Credentials(username, password));
 	}
 
 	public Responder wwwAuthenticate() {