+++ src/main/java/com/hencjo/summer/security/api/Authenticator2.java
@@ -1,7 +1,0 @@
-package com.hencjo.summer.security.api;
-
-import java.util.Optional;
-
-public interface Authenticator2 {
-	Optional<String> authenticate(String username, String password);
-}