+++ src/main/java/com/hencjo/summer/security/api/CredentialsAuthenticator.java
@@ -1,0 +1,17 @@
+package com.hencjo.summer.security.api;
+
+import java.util.Optional;
+
+public interface CredentialsAuthenticator {
+	Optional<String> authenticate(Credentials credentials);
+
+	final class Credentials {
+		public final String username;
+		public final String password;
+
+		public Credentials(String username, String password) {
+			this.username = username;
+			this.password = password;
+		}
+	}
+}